Revert "WIP implementation of counting word frequencies"
This reverts commit a02b637b57f54fcee63f2204a1c9cd94cb9a85aa.

This implementation slows down the performance by a huge margin. The
biggest culprit after a bit of profiling turns out to be higher GC
pressure caused by creation of strings in the frequency counting
routines (fmt.Sprintf calls). This defeats the purpose of hashing a
string since we are back to creating strings. Only difference between
just using those words as keys is that the length is now constant due to
hashing.

Another problem here is that hashing increases the size of the data
stored for smaller characters. For instance a 32bit fnv hash results in
a byte slice of length 4. This corresponds to a 4 character ASCII
string. So even characters like "a", "an", "the" etc will be hashed to a
string that is actually longer than the original ones. This is an
undesired side-effect. It's best not to create a string perhaps, and
just store the byte array in a DS. That DS should also then have a way
to check equality of a byte slice, and then sort the keys based on
values.